Test: デプロイテスト

+++ src/components/admin/bus/BusCategoryList.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { BusCategoryListType, ScheduleType } from "../../../types/admin";
-import { customAxios } from "../../../services/customAxios";
+import { privateApi } from "../../../services/customAxios";
 import { ListBtn } from "../../master/UserList";
 import BusScheduleList from "./BusScheduleList";
 
@@ -30,10 +30,17 @@ function BusCategoryList(props: BusCategoryListType) {
     setDropdown(false);
   }, [id]);
 
+  // 수정할 시
+  useEffect(() => {
+    if (modify) {
+      setNewName(round);
+    }
+  }, [modify]);
+
   // 스케줄 데이터 가져오기
   const getScheduleData = async () => {
     try {
-      const scheduleData = await customAxios.get(
+      const scheduleData = await privateApi.get(
         `/api/bus/round/schedule/${id}`
       );
       setSchedule(scheduleData.data.schedules);
@@ -45,7 +52,7 @@ function BusCategoryList(props: BusCategoryListType) {
   // 스케줄 생성하기
   const createSchedule = async (id: string, station: string, time: string) => {
     try {
-      await customAxios.post("/api/bus/schedule", {
+      await privateApi.post("/api/bus/schedule", {
         round_id: id,
         station: station,
         bus_time: time,
@@ -58,7 +65,7 @@ function BusCategoryList(props: BusCategoryListType) {
   // 스케줄 삭제하기
   const deleteSchedule = async (id: string) => {
     try {
-      await customAxios.delete(`/api/bus/schedule/${id}`);
+      await privateApi.delete(`/api/bus/schedule/${id}`);
     } catch (error) {
       console.log(error);
     }
@@ -79,7 +86,7 @@ function BusCategoryList(props: BusCategoryListType) {
             />
             <ListBtn
               value="완료"
-              color="bg-sky-400"
+              color="bg-sky-400/90"
               onClick={() => {
                 modifyCategotyFuc(id, newName).then(() => {
                   setModify(false);
@@ -93,7 +100,7 @@ function BusCategoryList(props: BusCategoryListType) {
             />
             <ListBtn
               value="취소"
-              color="bg-red-500/80"
+              color="bg-red-400/90"
               onClick={() => {
                 setModify(false);
               }}
@@ -104,22 +111,27 @@ function BusCategoryList(props: BusCategoryListType) {
             <div className="min-w-32 text-center">{round}</div>
             <ListBtn
               value="수정"
-              color="bg-pink-500/80"
+              color="bg-orange-400/80"
               onClick={() => {
                 setModify(true);
               }}
             />
             <ListBtn
               value="삭제"
-              color="bg-red-500/80"
+              color="bg-red-400/90"
               onClick={() => {
-                deleteCategoryFuc(id).then(() => {
-                  getCategoryFuc({
-                    semester: semester,
-                    weekend: weekend,
-                    bus_route_direction: bus_route_direction,
+                if (window.confirm("삭제하시겠습니까?")) {
+                  alert("삭제되었습니다");
+                  deleteCategoryFuc(id).then(() => {
+                    getCategoryFuc({
+                      semester: semester,
+                      weekend: weekend,
+                      bus_route_direction: bus_route_direction,
+                    });
                   });
-                });
+                } else {
+                  alert("취소되었습니다.");
+                }
               }}
             />
           </>
