+++ src/components/master/MasterMain.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { customAxios } from "../../services/customAxios";
+import { privateApi } from "../../services/customAxios";
 import { GetUserData } from "../../types/auth";
 import { ListHead, UserList } from "./UserList";
 import { GetApprovedData } from "../../types/master";
@@ -14,12 +14,12 @@ function MasterMain() {
   ];
   const dataList = [
     { value: "name", col: "col-span-1" },
-    { value: "phone_number", col: "col-span-1" },
+    { value: "phone_number", col: "col-span-1", type: "phoneNum" },
     { value: "email", col: "col-span-2" },
     [
       {
         value: "승인",
-        color: "bg-blue-400",
+        color: "bg-blue-400/90",
         onClick: (data: GetUserData) => {
           approval(data.id).then(() => {
             getData({ type: "unapproved" });
@@ -28,11 +28,16 @@ function MasterMain() {
       },
       {
         value: "거절",
-        color: "bg-red-400",
+        color: "bg-red-400/90",
         onClick: (data: GetUserData) => {
-          deleteData(data.id).then(() => {
-            getData({ type: "unapproved" });
-          });
+          if (window.confirm("거절하시겠습니까?")) {
+            alert("거절되었습니다");
+            deleteData(data.id).then(() => {
+              getData({ type: "unapproved" });
+            });
+          } else {
+            alert("취소되었습니다.");
+          }
         },
       },
     ],
@@ -48,7 +53,7 @@ function MasterMain() {
       const config: AxiosRequestConfig = {
         params: data,
       };
-      const unapprovedData = await customAxios.get("/api/admin/list", config);
+      const unapprovedData = await privateApi.get("/api/admin/list", config);
       setUnapprovedUsers(unapprovedData.data.admins);
     } catch (error) {
       console.log(error);
@@ -58,7 +63,7 @@ function MasterMain() {
   // 유저 데이터 승인요청
   const approval = async (data: number) => {
     try {
-      await customAxios.patch("/api/admin/approve", {
+      await privateApi.patch("/api/admin/approve", {
         admin_id: data,
         approve: 1,
       });
@@ -70,7 +75,7 @@ function MasterMain() {
   // 유저데이터 거절 삭제하기
   const deleteData = async (data: number) => {
     try {
-      await customAxios.delete(`/api/admin/master/${data}`);
+      await privateApi.delete(`/api/admin/master/${data}`);
     } catch (error) {
       console.log(error);
     }
